Rename file, style changes,  wrk as default
